requirements.txt
streamlit
openai
pinecone-client
sentence-transformers
pandas
torch